InfiniteScroller: Improve scrollParent detection + add onScroll callback
